Peer Code Review
This commit is full of CSS changes and some minor HTML tweaks. All
images are saved out, but not currently within the design, and type is
being changed to the correct sizes. Lots of padding and margins has
been added to some sections to give the design the correct look, and
nothing has been done on small display other than having the correct
structure.