Assignment Six
Last commit for this assignment. This has the added code from Elizabeth
to make the product section work correctly, selected product images
correctly to have those displayed, got the responsiveness for the
product buttons to work, and cleaned CSS to somewhat of a better state
than what it was in.